requirements.txt
python-dotenv
requests
firecrawl
openai
